+++ packages/ak-editor-toolbar-hyperlink/src/index.js
@@ -1,6 +1,4 @@
-/** @jsx vdom */
 import 'style!./host.less';
-
 import { vdom, define, prop, emit, symbols } from 'skatejs';
 import shadowStyles from './shadow.less';
 import EditorButton from 'ak-editor-button';
@@ -8,10 +6,26 @@ import Icon from 'ak-editor-icon';
 import Popup from 'ak-editor-popup';
 import TextInput from 'ak-editor-popup-text-input';
 
-const definition = {
+function toggle(elem, input) {
+  elem.open = !elem.open;
+
+  if (elem.open) {
+    const textInput = input || elem[symbols.shadowRoot].querySelector('.text-input');
+
+    // todo: fix the hack
+    setTimeout(() => textInput.focus(), 5);
+  }
+}
+
+export default define('ak-editor-toolbar-hyperlink', {
   render(elem) {
     const LinkButton = (<EditorButton
-      class="link-button"
+      className="link-button"
+      onClick={() => {
+        if (!elem.disabled) {
+          toggle(elem);
+        }
+      }}
       disabled={elem.disabled}
     ><Icon glyph="link" /></EditorButton>);
 
@@ -21,19 +35,7 @@ const definition = {
     /* eslint-disable new-cap  */
     return (
       <div
-        onclick={() => {
-          if (!elem.disabled) {
-            const popup = elem[symbols.shadowRoot].querySelector('.popup');
-            popup.open = true;
-
-            const textInput = elem[symbols.shadowRoot].querySelector('.text-input');
-
-            // todo: fix the hack
-            setTimeout(() => textInput.focus(), 5);
-          }
-        }}
-
-        onkeyup={event => {
+        onKeyup={event => {
           if (event.keyCode === 13) {
             const textInput = elem[symbols.shadowRoot].querySelector('.text-input');
             const popup = elem[symbols.shadowRoot].querySelector('.popup');
@@ -49,14 +51,16 @@ const definition = {
         <Popup
           class="popup"
           target={linkButton}
+          open={elem.open}
+          on-ak-blanket-click={() => toggle(elem)}
         >
-          <TextInput class="text-input" placeholder="Paste link" />
+          <TextInput className="text-input" placeholder="Paste link" />
         </Popup>
       </div>
     );
   },
   props: {
-    /* eslint-disable max-len  */
+    disabled: prop.boolean({ attribute: true }),
     /**
      * @description Controls disablily of an popup.
      * @memberof Popup
@@ -66,11 +70,6 @@ const definition = {
      * @example @html <ak-editor-popup disabled></ak-editor-popup>
      * @example @js dialog.disabled = true;
      */
-    disabled: prop.boolean({
-      attribute: true,
-      default: false,
-    }),
+    open: prop.boolean({ attribute: true }),
   },
-};
-
-export default define('ak-editor-toolbar-hyperlink', definition);
+});