+++ packages/akutil-polyfills/src/index.js
@@ -1,5 +1,2 @@
-import 'es6-map/implement'; // for Android 4.0
-import 'webcomponents.js/src/WeakMap/WeakMap.js';  // for Android 4.0
-import 'webcomponents.js/src/MutationObserver/MutationObserver.js';  // for Android 4.0
-import 'skatejs-named-slots';
-import 'webcomponents.js/src/CustomElements/v1/CustomElements.js';
+import 'es6-symbol/implement';
+import 'skatejs-web-components';