+++ packages/pf-mention/src/api/pf-mention-resource.js
@@ -1,7 +1,7 @@
 import URLSearchParams from 'url-search-params'; // IE, Safari, Mobile Chrome, Mobile Safari
 require('es6-promise').polyfill();
 import 'whatwg-fetch';
-
+import debug from '../util/logger';
 
 const buildUrl = (baseUrl, path, data, secOptions) => {
   const searchParam = new URLSearchParams();
@@ -72,22 +72,22 @@ const requestService = (baseUrl, path, data, secOptions) => {
 class AbstractMentionResource {
 
   constructor() {
-    this._changeListeners = [];
-    this._errListeners = [];
+    this._changeListeners = new Map();
+    this._errListeners = new Map();
   }
 
-  subscribe(callback, errCallback) {
+  subscribe(key, callback, errCallback) {
     if (callback) {
-      this._changeListeners.push(callback);
+      this._changeListeners.set(key, callback);
     }
     if (errCallback) {
-      this._errListeners.push(errCallback);
+      this._errListeners.set(key, errCallback);
     }
   }
 
-  unsubscribe(callback, errCallback) {
-    this._changeListeners = this._changeListeners.filter((v) => v !== callback);
-    this._errListeners = this._errListeners.filter((v) => v !== errCallback);
+  unsubscribe(key) {
+    this._changeListeners.delete(key);
+    this._errListeners.delete(key);
   }
 
   filter(query) {
@@ -95,27 +95,27 @@ class AbstractMentionResource {
   }
 
   _notifyListeners(mentions) {
-    this._changeListeners.forEach((listener) => {
+    debug('pf-mention-resource._notifyListeners',
+      mentions && mentions.mentions && mentions.mentions.length,
+      this._changeListeners);
+
+    this._changeListeners.forEach((listener, key) => {
       try {
         listener(mentions.mentions);
       } catch (e) {
         // ignore error from listener
-        if (process.env.NODE_ENV === 'development') {
-          console.log('error from listener, ignoring', e); // eslint-disable-line no-console
-        }
+        debug(`error from listener '${key}', ignoring`, e);
       }
     });
   }
 
   _notifyErrorListeners(error) {
-    this._errListeners.forEach((listener) => {
+    this._errListeners.forEach((listener, key) => {
       try {
         listener(error);
       } catch (e) {
         // ignore error from listener
-        if (process.env.NODE_ENV === 'development') {
-          console.log('error from listener, ignoring', e); // eslint-disable-line no-console
-        }
+        debug(`error from listener '${key}', ignoring`, e);
       }
     });
   }
@@ -152,7 +152,6 @@ class MentionResource extends AbstractMentionResource {
    *          key2: value,
    *        }
    *      }
-   *   application: string (required), // the site, or product
    *   containerId: string (optional)
    * }
    *
@@ -166,11 +165,7 @@ class MentionResource extends AbstractMentionResource {
     if (!config.securityProvider) {
       throw new Error('config.securityProvider is a required parameter');
     }
-    if (!config.application) {
-      throw new Error('config.application is a required parameter');
-    }
 
-    // FIXME clone/immutable js?
     this._config = config;
     this._lastReturnedSearch = 0;
   }
@@ -182,10 +177,8 @@ class MentionResource extends AbstractMentionResource {
         this._lastReturnedSearch = searchTime;
         this._notifyListeners(mentions);
       } else {
-        if (process.env.NODE_ENV === 'development') {
-          const date = new Date(searchTime).toISOString().substr(17, 6);
-          console.log('Stale search result, skipping', date, query); // eslint-disable-line no-console, max-len
-        }
+        const date = new Date(searchTime).toISOString().substr(17, 6);
+        debug('Stale search result, skipping', date, query); // eslint-disable-line no-console, max-len
       }
     };
 
@@ -209,9 +202,7 @@ class MentionResource extends AbstractMentionResource {
    */
   _initialState() {
     const secOptions = this._config.securityProvider();
-    const options = {
-      application: this._config.application,
-    };
+    const options = {};
 
     if (this._config.containerId) {
       options.containerId = this._config.containerId;
@@ -223,7 +214,6 @@ class MentionResource extends AbstractMentionResource {
     const secOptions = this._config.securityProvider();
     const options = {
       query,
-      application: this._config.application,
     };
     if (this._config.containerId) {
       options.containerId = this._config.containerId;