+++ packages/ak-editor-toolbar/CHANGELOG.md
@@ -1,5 +1,25 @@
-<a name="2.0.0"></a>
-# 2.0.0 (2016-08-01)
+<a name="2.0.2"></a>
+## 2.0.2 (2016-08-19)
+
+
+
+<a name="ak-editor-toolbar@2.0.2"></a>
+## ak-editor-toolbar@2.0.2 (2016-08-08)
+
+
+### Bug Fixes
+
+* Disable ::slotted(*) styling until a fix to avoid rule merging is found. ([0e835c3](https://bitbucket.org/atlassian/atlaskit/commits/0e835c3))
+
+
+
+<a name="ak-editor-toolbar@2.0.1"></a>
+## ak-editor-toolbar@2.0.1 (2016-08-04)
+
+
+### Bug Fixes
+
+* **dependencies:** Dependency to ak-blanket was on 0.0.0 ([fc4a7aa](https://bitbucket.org/atlassian/atlaskit/commits/fc4a7aa))
 
 
 
@@ -9,11 +29,37 @@
 
 ### Bug Fixes
 
-* endsWith is not available in IE ([d8b1f01](https://bitbucket.org/atlassian/atlaskit/commits/d8b1f01))
-* FAB-370 add attributes to prop types in pf-mention-picker ([b1929da](https://bitbucket.org/atlassian/atlaskit/commits/b1929da))
-* Fix bad export for MentionResource ([1502f73](https://bitbucket.org/atlassian/atlaskit/commits/1502f73))
-* Remove broken export from ak-inline-dialog, and prevent usage upstream. ([e5cc55f](https://bitbucket.org/atlassian/atlaskit/commits/e5cc55f))
 * **package:** Add ak- prefix to editor-*, and fix licensing. ([15751dc](https://bitbucket.org/atlassian/atlaskit/commits/15751dc))
+
+
+
+<a name="0.0.1-atlas-animation3"></a>
+## 0.0.1-atlas-animation3 (2016-06-08)
+
+
+
+<a name="1.1.6-atlas-animation"></a>
+## 1.1.6-atlas-animation (2016-06-08)
+
+
+
+<a name="13.1.0-tmp-atlaskit-component-b"></a>
+# 13.1.0-tmp-atlaskit-component-b (2016-06-08)
+
+
+
+<a name="13.1.0-tmp-atlaskit-component-a"></a>
+# 13.1.0-tmp-atlaskit-component-a (2016-06-08)
+
+
+### Bug Fixes
+
+* **components:** commented out a line ([4b91174](https://bitbucket.org/atlassian/atlaskit/commits/4b91174))
+
+
+### Features
+
+* **components:** Add an ! mark ([2b7bdb0](https://bitbucket.org/atlassian/atlaskit/commits/2b7bdb0))